clean up download links to be included and use version const vars
now to update release versions just change the one release version
variable in download-latest.txt at the top. if a new library, app etc.
gets added, add it to the table given the const variable setup similar
to what is already there and add a new version variable for it at the
top.