www: eo-docs: initial commit of our generated eo based API documentation
Generated with the gendoc lua script to parse the eo file documentation and
generate files matching the dokuwiki syntax. This will be integrated into our
main website dokuwiki setup at some point. For now we keep the files separate
to allow for more testing and potential failures.

At the current state the documentation covers 100% of the item we have in our
eo files. If you want to play around with this on your own I would recommend a
local setup as described here:

https://phab.enlightenment.org/w/run_doc_system_locally/